requirements.txt
imageio==2.33.0
Pillow==10.1.0
Pillow==10.1.0
